pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdftools"
version = "1.0.0"
description = "A comprehensive Python toolkit for PDF processing with multiple OCR backends"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "PDFTools Contributors"}
]
requires-python = ">=3.8"
keywords = ["pdf", "ocr", "mistral", "text-extraction", "document-processing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]

dependencies = [
    "mistralai>=0.1.0",
    "tenacity>=8.2.0",
    "streamlit>=1.28.0",
    "watchdog>=3.0.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
    "ruff>=0.10.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
]

docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/PDFTools"
Documentation = "https://github.com/yourusername/PDFTools#readme"
Repository = "https://github.com/yourusername/PDFTools.git"
Issues = "https://github.com/yourusername/PDFTools/issues"

[project.scripts]
pdftools-ocr = "pdftools.mistral_ocr:main"

[tool.setuptools]
packages = ["pdftools"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_glob = ["*/migrations/*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov=pdftools --cov-report=html --cov-report=term-missing"

[tool.coverage.run]
source = ["pdftools"]
omit = [
    "*/tests/*",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.ruff.isort]
profile = "black"
